logic-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
vaderSentiment==3.3.2
spotipy==2.23.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
